+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [Unreleased]
+
+- Trigger disconnect when entities with `ConnectedClient` are despawned.
+
 ## 0.9.1 - 04/27/25
 
 - Fix compile error with `bevy` Resource derives.